+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added I2C and SPI APIs to rp2 platform
 - Added `code:get_object_code/1`
 - Added `erlang:display_string/1` and `erlang:display_string/2`
+- Added Thumb-2 support to armv6m JIT backend, optimizing code for ARMv7-M and later cores
 
 ### Changed
 - ~10% binary size reduction by rewriting module loading logic